Invent the future, don't recreate the past - FutureJS
I have seem I push with this talk but not merge.. so I guess it wasn't in the requirements to merge.

This is why this talk should definitely be in the js must watch.

"Compared to many modern languages, JavaScript is somewhat minimal. It has first-class functions, but a supply of functional tools. It has objects, but its inheritance model is sparse and its metaobjects are hacked on top of functions. This is frustrating to some and a blessing to others. It does not force a heavyweight paradigm upon you, and if you need more you can build it yourself. And everybody does. But all too-often, we look backwards and re-create past, when we could be looking forward and inventing future.

Instead of copying ideas from 1984, we should invent ideas that others will copy in 2044."