Add Hey Underscore, You're Doing It Wrong!
This video talks about functional programming basics such as currying
and composing and how common libraries like underscore and lodash
prevents their users from utilizing these capabilities. It is a very
good and often mindblowing introduction of functional programming to
people that code JavaScript but are not aware of the funtional
capabilities the language has.