Doug Crockford: JavaScript: The Good Parts
This talk a a classic, summarizing what is bad with JS (semicolon insertion, type coercion, ...) that we should avoid and what is good that we should leverage (lambdas, dynamic objects, ...). The importance of using something like JSHint is stressed. It is in a way elementary as all experienced JS devs should know these things but people that use the language without ever learning it properly - like me - may find it enlightening.